refactor(pyo3): extract gen_typeddict to a sibling module
Blast-radius review follow-up: the Python field-typing additions pushed types.rs over the 800-line
split-before-adding guideline. Move the self-contained TypedDict generator into types/typeddict.rs
(728 lines remaining). Pure move + comment tightening — generated output is byte-identical.
